+++ docs/news.rst
@@ -14,6 +14,7 @@ Python 2.6 is no longer supported by scrapyd.
 Added
 ~~~~~
 
+- Make spider queue class configurable
 - Include run's PID in listjobs webservice.
 - Include full tracebacks from scrapy when failing to get spider list.
   This will lead to more noisy webservice output